refactor: move blockchain tree traits to own crate
