+++ crates/blockchain-tree-api/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "reth-blockchain-tree-api"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+# reth
+reth-interfaces.workspace = true
+reth-primitives.workspace = true
+reth-rpc-types.workspace = true
