+++ crates/blockchain-tree/Cargo.toml
@@ -12,6 +12,7 @@ workspace = true
 
 [dependencies]
 # reth
+reth-blockchain-tree-api.workspace = true
 reth-primitives.workspace = true
 reth-interfaces.workspace = true
 reth-db.workspace = true
@@ -37,7 +38,7 @@ linked_hash_set = "0.1.4"
 [dev-dependencies]
 reth-db = { workspace = true, features = ["test-utils"] }
 reth-interfaces = { workspace = true, features = ["test-utils"] }
-reth-primitives = { workspace = true , features = ["test-utils"] }
+reth-primitives = { workspace = true, features = ["test-utils"] }
 reth-provider = { workspace = true, features = ["test-utils"] }
 reth-revm.workspace = true
 reth-node-ethereum.workspace = true
@@ -46,4 +47,9 @@ assert_matches.workspace = true
 
 [features]
 test-utils = []
-optimism = ["reth-primitives/optimism", "reth-interfaces/optimism", "reth-provider/optimism", "reth-revm/optimism"]
+optimism = [
+    "reth-primitives/optimism",
+    "reth-interfaces/optimism",
+    "reth-provider/optimism",
+    "reth-revm/optimism",
+]
