+++ crates/blockchain-tree/src/blockchain_tree.rs
@@ -5,11 +5,13 @@ use crate::{
     state::{BlockChainId, TreeState},
     AppendableChain, BlockIndices, BlockchainTreeConfig, BundleStateData, TreeExternals,
 };
+use reth_blockchain_tree_api::{
+    BlockAttachment, BlockStatus, BlockValidationKind, CanonicalOutcome, InsertPayloadOk,
+};
 use reth_db::database::Database;
 use reth_interfaces::{
     blockchain_tree::{
-        error::{BlockchainTreeError, CanonicalError, InsertBlockError, InsertBlockErrorKind},
-        BlockAttachment, BlockStatus, BlockValidationKind, CanonicalOutcome, InsertPayloadOk,
+        BlockchainTreeError, CanonicalError, InsertBlockError, InsertBlockErrorKind,
     },
     consensus::{Consensus, ConsensusError},
     executor::{BlockExecutionError, BlockValidationError},
