+++ crates/consensus/beacon/Cargo.toml
@@ -12,6 +12,7 @@ workspace = true
 
 [dependencies]
 # reth
+reth-blockchain-tree-api.workspace = true
 reth-beacon-consensus-core.workspace = true
 reth-primitives.workspace = true
 reth-interfaces.workspace = true
