+++ crates/interfaces/src/lib.rs
@@ -12,6 +12,9 @@
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
+/// Blockchain tree errors.
+pub mod blockchain_tree;
+
 /// Consensus traits.
 pub mod consensus;
 
@@ -36,9 +39,6 @@ pub mod provider;
 
 /// Syncing related traits.
 pub mod sync;
-
-/// BlockchainTree related traits.
-pub mod blockchain_tree;
 
 #[cfg(any(test, feature = "test-utils"))]
 /// Common test helpers for mocking out Consensus, Downloaders and Header Clients.
